+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
-## [Unreleased]
+## [1.2.0](https://marketplace.visualstudio.com/_apis/public/gallery/publishers/shoreless/vsextensions/symfony-translation-helper/1.0.0/vspackage)
 
-- Minor/bugfix releases for internal evaluation.  
+- Add translation domain extraction to language file parsers.  
+- Show translation domain in translations popups.  
+
+## [1.1.0](https://marketplace.visualstudio.com/_apis/public/gallery/publishers/shoreless/vsextensions/symfony-translation-helper/1.0.0/vspackage)
+
+- Speed up opening of translation files from within translations popups.  
+- Open translation file with selected corresponding translation string.
+
+## [1.0.0](https://marketplace.visualstudio.com/_apis/public/gallery/publishers/shoreless/vsextensions/symfony-translation-helper/1.0.0/vspackage)
+
+- Initial public release.